Move utils.py to core/pipelines.py
