+++ orchest/orchest-api/app/app/apis/namespace_runs.py
@@ -4,10 +4,10 @@ from celery.task.control import revoke
 from flask import current_app, request
 from flask_restplus import Namespace, Resource, fields
 
-from app.connections import db
-import app.models as models
 from app.celery_app import make_celery
-from app.utils import construct_pipeline
+from app.connections import db
+from app.core.pipelines import construct_pipeline
+import app.models as models
 
 
 api = Namespace('runs', description='Managing (partial) runs')
