+++ orchest/orchest-api/app/app/core/tasks.py
@@ -6,7 +6,7 @@ from celery import Task
 
 from app import create_app
 from app.celery_app import make_celery
-from app.utils import Pipeline, PipelineDescription
+from app.core.pipelines import Pipeline, PipelineDescription
 from config import CONFIG_CLASS
